God help us

+++ Source/SteamHub.ApiContract/Models/User/User.cs
@@ -44,9 +44,9 @@ namespace SteamHub.ApiContract.Models.User
 
         public string Email { get; set; }
 
-        public float WalletBalance { get; set; }
+        public decimal WalletBalance { get; set; }
 
-        public float PointsBalance { get; set; }
+        public int PointsBalance { get; set; }
 
         public UserRole UserRole { get; set; }
